Btrfs: use WB_SYNC_ALL for sync ioctl and snapshot creation sync
When we create a snapshot, we write out all of the dirty inodes in that
subvolume to make sure the snapshot matches the source.  This goes
through a twisty path into a delalloc helper function that is also used
to send down inodes for ENOSPC accounting.

These two uses have different needs, one is similar to sync and the
other should be just like background writeback.  The code currently uses
filemap_flush() which isn't really suitable for sync.

This changes things to pass in a sync_mode and pick the correct flushing
function based on the mode.  The btrfs filesystem sync ioctl uses the
same code as snapshot creation, so this fixes up all of them at once.

This isn't relevant upstream unilt our skip-the-last-page patches
go upstream.

Tasks: 20253127

Signed-off-by: Chris Mason <clm@fb.com>