btrfs: handle async pages properly with delalloc flushing
Summary:
btrfs: handle async pages properly with delalloc flushing

sync_inode() will simply skip inodes that are not marked dirty and have
WB_SYNC_NONE set.  This means that if we wrote the full inode with
compression the subsequent sync_inode() won't do anything.  We had code
to deal with this before, but I pulled it out because I thought it
wasn't needed.  Fix this by putting back the old code which waits for
the async pages to settle before doing any of the other waiting.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <jbacik@fb.com>
Signed-off-by: Chris Mason <clm@fb.com>