btrfs: just commit the transaction
Summary:
btrfs: just commit the transaction

We have all of this logic to keep us from committing the transaction if
we don't need to, but it is from when we could commit the transaction
all the time and not make progress.  We don't need this logic anymore,
just commit the transaction if we think it'll help.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Signed-off-by: Josef Bacik <jbacik@fb.com>
Signed-off-by: Chris Mason <clm@fb.com>