string_helpers: Readd *_escape_mem_ascii()
Summary:
string_helpers: Readd *_escape_mem_ascii()

This is needed for 5.6 seq_file to function with a newer string_helpers,
since this was removed in the interim due to API changes.

This partially reverts commit 815c913a0dc94b75c4b5bcb850817d912f5d9400,
which isn't in this tree, but was part of the big string_helpers
backport.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags: