string_helpers: Backport from v5.13-rc6
Summary:

This is a self-contained backport of string_helpers and its tests from
v5.13-rc6, which the printk patches are based on. This is required to support
ESCAPE_APPEND (which we use later), and set the base for the new quoting
behaviour in commit "string_helpers: Escape double quotes in escape_special".

Test Plan:

Tested as part of larger merge.

Reviewers: jweiner, alastorze, kernel_ops

Tasks: T93369608