btrfs: set wbc->reason properly when ENOSPC flushing
Because of our O_APPEND stuff we will skip writing if we do WB_SYNC_NONE
and we're in the last page of the inode.  In order to make sure we
actually write we need to have wbc->reason set to
WB_REASON_FS_FREE_SPACE.  Our fix for flushing missed this fact and so we
could skip writing out some inodes and thus cause early enospc's.  Fix
this by noting that this is for an ENOSPC flush and set wbc->reason
properly.

Signed-off-by: Chris Mason <clm@fb.com>